spi/mpc8xxx: refactor the common code for SPI/eSPI controller
commit b36ece832512c1a0afa54ff0a56d63492a1caf08 upstream
correspond to vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0108-spi-mpc8xxx-refactor-the-common-code-for-SPI-eSPI-co.patch

Refactor the common code in file spi_mpc8xxx.c to spi_fsl_lib.c
used by SPI/eSPI controller driver as a library, move the SPI
controller driver to a new file spi_fsl_spi.c, and leave the
QE/CPM SPI controller code in this file.

Because the register map of the SPI controller and eSPI controller
is so different, also leave the code operated the register to the
driver code, not the common code.

Signed-off-by: Mingkai Hu <Mingkai.hu@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>