net: Allow skb_recycle_check to be done in stages
Extracted from vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0147-net-Allow-skb_recycle_check-to-be-done-in-stages.patch

skb_recycle_check resets the skb if it's eligible for recycling.
However, there are times when a driver might want to optionally
manipulate the skb data with the skb before resetting the skb,
but after it has determined eligibility.  We do this by splitting the
eligibility check from the skb reset, creating two inline functions to
accomplish that task.

Signed-off-by: Andy Fleming <afleming@freescale.com>