tracing: fix oops for powerpc in irqsoff_tracer
When CONFIG_IRQSOFF_TRACER enabled, oops will be seen in lttng test
and strace.

The root cause is that only privilege mode is taken into account to
get CALLER_ADDR1 in trace_hardirqs_on by toolchain, and it caters
for IRQSOFF_TRACER.

So this fix will passby IRQSOFF_TRACER when restore to user-mode.

Signed-off-by: Liming Wang <liming.wnag@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>