x86: save bp twice to set bp field in pt_regs properly
in commit 097a0788df71b0f3328c70ab5f4e41c27ee66817, the bp field
has been saved correctly.

but in commit d99015b1abbad743aa049b439c1e1dede6d0fa49, the bp field
was removed by mistake.

So redo what the first commit does:

Save rbp twice: One is for marking the stack frame, as usual (already
there), and the other, to fill pt_regs properly. This is because bx
comes right before the last saved register in that structure, and not
bp. If the base pointer were in the place bx is today, this would not
be needed.

Signed-off-by: Liming Wang <liming.wang@windriver.com>