sctp: fix oops with unconnected sockets
In a few boundary cases a socket could be used where it isn't in a
connected or listening state for SCTP traffic.  When this occurs
the attached network is NULL which in-turn causes the __xfrm_lookup()
function to panic by dereferencing the NULL net.

To address this a check is added in the SCTP get_dst() function where
if a net attached to a socket is NULL, the function simply returns
NULL indicating no destination for the intended SCTP association as
the socket itself has no connection.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>