tracing/ftrace: MIPS: Fixup of the support for SMP
To flush the icaches on the other cpus, the irqs of the cpus should be
enabled to accept the ipi interrupts. otherwise, the icaches on the
other cpus may not be flushed and even make the system hang: the current
cpu may wait for the other cpus to flush their icaches, but the other
cpus can not response for the irqs are disabled.

To fix this problem, this patch delays the icache flushing operation
after the irqs are enabled.

Benefit from this fix, the system performance will be improved for
currently, we only need to flush the whole icache one time at the end
instead of flush thousands of entries individually.

A 'side effect' is, from now on, any user of ftrace_modify_code() in
MIPS should flush the icache itself.

The *command may be the OR result of the following values:

include/linux/ftrace.h:

	FTRACE_ENABLE_CALLS		= (1 << 0),
	FTRACE_DISABLE_CALLS		= (1 << 1),
	FTRACE_UPDATE_TRACE_FUNC	= (1 << 2),
	FTRACE_ENABLE_MCOUNT		= (1 << 3),
	FTRACE_DISABLE_MCOUNT		= (1 << 4),
	FTRACE_START_FUNC_RET		= (1 << 5),
	FTRACE_STOP_FUNC_RET		= (1 << 6),

If the FTRACE_ENABLE_CALLS or FTRACE_DISABLE_CALLS, we must flush the
whole icache for we don't know which kernel entries are modified. if
only one of the others is set, we just need to flush the related entry.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>