MIPS: disable subsection optimization in atomic_*
The current implementation of atomic_* uses .subsection directives to
move the unlikely condition out of line, in a similar way that spinlocks
do.  But there is an assumption that the out of line code won't be moved
more than 128k away.

on big files like alloc.c with sizes like this:

    ocfs2$ size alloc.o
       text        data     bss     dec     hex filename
     137304         400       0  137704   219e8 alloc.o
    ocfs2$

you break the 128k limit and get these errors

    {standard input}: Assembler messages:
    {standard input}:871: Error: Branch out of range

And given that Ralf's tests showed there was no super awesome gain from
the subsections, So, disable the subsection optimization and go back to
the simple method as the R10000_LLSC_WAR does.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>