tracing: don't trace trace-clock
trace-clock is used for tracing, to reduce garbage information from the
tracing result, we'd better not trace it.

FYI:

"CFLAGS_REMOVE_trace-clock.o = -pg" prevents the compiler from compiling
the file with -pg and as a result, all of the functions in the related
files will not be traced.

"notrace" tells the compiler not to trace an indicated function.

functions with "_notrace" suffix is the notrace variants of the
functions, will not be traced by the preemptoff tracer.

the "raw_" variants of local_irq_* will not be traced by the irqsoff
tracer.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>