tracing: don't trace ftrace
We replace the callees of ftrace by the notrace variants to reduce more
garbage information from the tracing result.

FYI:

functions with "_notrace" suffix is the notrace variants of the
functions, will not be traced by the preemptoff tracer.

the "raw_" variants of local_irq_* will not be traced by the irqsoff
tracer.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>