kexec/ppc32: Pass irq_regs to crash IPI
kexec handling of crash kernel uses IPIs to signal the other CPUs
that they should stop executing whatever they were doing. It's
possible that the IPI implementation uses doorbell interrupts. If
that is the case, the irq_regs are not passed to the IPI handler,
causing a possible dereference of a NULL pointer.

Add passing of irq_regs, similar to what do_IRQ() implements.

Signed-off-by: Benjamin Walsh <benjamin.walsh@windriver.com>
[context adjustment]
Integrated-by: Weiwei Wang <weiwei.wang@windriver.com>