kexec/ppc32: purgatory expect r3 to contain boot CPU#
Purgatory will get the value of r3 and force it in the DT
passed to the second kernel. If the boot CPU is not the same
as that value, the boot of the second kernel will be
interrupted and the kernel will oops.

Hmm, how come we've never been bitten by this one ? It only
caused a problem on P4080 crash reboot.

Signed-off-by: Benjamin Walsh <benjamin.walsh@windriver.com>