tracing/trace-clock: Remove tracing_ctrl_release
The commit "tracing/ftrace: Fixes the trace_clock_local() for the boards
who use the jiffies based sched_clock()" used tracing_ctrl_release() to
release the trace clock for Lttng, but this function will be called when
tracing/tracing_enabled is closed, after enabling the tracer with the
following command:

$ echo 1 > /path/to/tracing/tracing_enabled

The trace clock will be released, as a result, the tracing will get the
low resolution timestamp and the user-space Ftrace will not benefit from
trace-clock-32-to-64.

For the selftests, because they don't open and close
tracing/tracing_enabled, so, tracing_ctrl_release() will not be executed
and the trace clock will not be released, so, no problem with selftests.

To release the trace clock for Lttng, we have only one method, that is,
switch to the nop tracer to remove all of the tracers from tracing, it
will unregister all of the tracers, therefore the
unregister_trace_clock() will be called and the trace clock will be
released.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>