mpc512x_dma: fix the hanged transfer issue
Current code clears interrupt active status _after_ submiting new
transfers. This leaves a possibility of clearing the interrupt for this
new transfer (if it is triggered fast enough) and thus lose this
interrupt. We want to clear interrupt active status _before_ new
transfers is submited and for current channel only.

<This is merged from upstream:
http://lists.ozlabs.org/pipermail/linuxppc-dev/2010-September/086084.html>

Signed-off-by: Ilya Yanok <yanok@emcraft.com>
Cc: Piotr Ziecik <kosmo@semihalf.com>
Integrated-by: Tonyliu <Bo.Liu@windriver.com>