FS Compat: Initialize iov pointer to prevent kernel oops
Fixes kernel oops after failing parameter checks.

The "iov" pointer is not initiatized prior to calling
compat_rw_copy_check_uvector such that a kernel oops is caused
when a parameter check failure jumps to the "unwind" code and kfree
is called.  The kernel oops looks like this:

    ...
    CPU 0 Unable to handle kernel paging request at virtual address
    000000af8176a4a0, epc == ffffffff81210af4, ra == ffffffff8126561c

    Oops[#17]:
    ... etc ...

    Call Trace:
    [<ffffffff81210af4>] kfree+0x5c/0x1b0
    [<ffffffff8126561c>] compat_do_readv_writev+0x54/0x208
    [<ffffffff81265ae0>] compat_sys_readv+0x68/0xd8
    [<ffffffff81103684>] handle_sysn32+0x44/0x84
    ...

kfree checks for NULL parameter so it is safe just to set iov
to NULL.

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>