powerpc/ppc64e: Fix link problem when building ppc64e_defconfig
commit 5b8544c38e6fde6968645afd46ff681492192b86 upstream
correspond to vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0050-powerpc-ppc64e-Fix-link-problem-when-building-ppc64e.patch

arch/powerpc/platforms/built-in.o:(.toc1+0x18): undefined reference to `__early_start'

This is due to the 85xx/smp.c not handling the 64-bit side properly.  We
need to set the entry point for secondary cores on ppc64e to
generic_secondary_smp_init instead of __early_start that we due on ppc32.

Signed-off-by: Kumar Gala <galak@kernel.crashing.org>