tipc: clean out all instances of #if 0'd unused code
commit 6ab8dc59b21ebd941993d8b71644cf952571b880 from
git://tipc.cslab.ericsson.net/pub/git/people/allan/tipc.git

Remove all instances of legacy or proposed-but-not-implemented code
that lives within an #if 0 ... #endif block. If some of it is needed
in the future it can recovered out of history, but there is no need
for it to clutter up the current code base.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <hao.xu@windriver.com>