tipc: Ensure both nodes recognize loss of contact between them
commit 0e2824ace97f11685ab7869c17f621e0c4d2c5ca from
git://tipc.cslab.ericsson.net/pub/git/people/allan/tipc.git

Enhances TIPC to ensure that a node that loses contact with a
neighboring node does not allow contact to be re-established until
it knows that its peer has also recognized the loss of contact.

Previously, nodes that were connected by two (or more) links could
encounter a situation in which node A would lose contact with node B
on all of its links, purge its name table of names published by B,
and then fail to repopulate those names once contact with B was restored;
this would happen because B was able to re-establish one or more links
so quickly that it never reached a point where it had no links to A --
meaning that B never saw a loss of contact with A, and consequently
didn't re-publish its names to A.

This problem is now prevented by enhancing the clean up done by TIPC
following a loss of contact with a neighboring node to ensure that
node A ignores all messages sent by B until it receives a LINK_PROTOCOL
message that indicates B has lost contact with A, thereby preventing
the (re)establishment of links between the nodes. The loss of contact
is recognized when a RESET or ACTIVATE message is received that has
a "redundant link exists" field of 0, indicating that B's sending link
endpoint is in a reset state and that B has no other working links.

Note: The reworking of the existing clean up code also eliminates a
check that prevents a link endpoint's discovery object from responding
to incoming messages while stale name table entries are being purged.
This check was unnecessary and potentially slowed down re-establishment
of communication between the nodes in some situations.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <hao.xu@windriver.com>