tipc: Clean out unused prototype for multiple discovery objects
commit 2574f9825b598a391a81f52aaa17d1d594a1a79a from
git://tipc.cslab.ericsson.net/pub/git/people/allan/tipc.git

Removes prototype code that was intended to allow a TIPC bearer
to have multiple discovery objects. This capability would have
allowed the bearer to have greater control over the nodes it
established links to, by allowing it to specify a subset specific
subset of nodes within a cluster (or clusters within a zone, or
zones within a network). If the prototype is needed in the future
it can recovered out of history, but there is no need for it to
clutter up the current code base.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <hao.xu@windriver.com>