tipc: Enhance handling of discovery object creation errors
commit b5f751575d831afcd108c3b6f06f674f682d6bf2 from
git://tipc.cslab.ericsson.net/pub/git/people/allan/tipc.git

Modifies TIPC's bearer code to better handle errors that can occur
when a new bearer is enabled. Changes include:

1) Now rejects attempts to enable a bearer whose discovery domain
   is its own node. (Such a bearer would be useless, since there
   is no way to change or extend the discovery domain.)

2) Now handles cases where the new bearer's discovery object can't
   be created. (Previously the bearer would be enabled, but would
   be unable to establish any links; now the bearer isn't enabled
   at all.)

The above changes allow some additional simplications to be made
to the bearer code that do not affect its existing operation.
This includes the removal of an unused return value when a bearer
is disabled, and the consolidation of two related routines that are
always called in tandem from a single location.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <hao.xu@windriver.com>