tipc: Prevent invalid memory access when sending to configuration service
commit 9139d38794c1c5c2e6b825d014f0bebb6163d278 from
http://tipc.cslab.ericsson.net/cgi-bin/gitweb.cgi?p=people/allan/tipc.git;a=shortlog;h=tipc1.7

Enhances network administration privilege checking when sending
a message to TIPC's configuration service so that it rejects messages
without a full message header. Previously, an application that sent a
message to the configuration service that was too short could cause
the validation code to access an uninitialized field in the msghdr
structure, resulting in a memory access exception.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <Hao.Xu@windriver.com>