tipc: Fix problem with missing link in "tipc-config -l" output
commit 252d24abc8bf6a7966523c7127da6ccbab467717 from
http://tipc.cslab.ericsson.net/cgi-bin/gitweb.cgi?p=people/allan/tipc.git;a=shortlog;h=tipc1.7

Removes a race condition that could cause TIPC's internal counter
of the number of link it has to neighboring nodes to have the
incorrect value if new links were simultaneously created on different
bearers. This under counting would result in TIPC failing to list
the final link(s) in it's response to a configuration request to
list all of the nodes links. The counter is now updated atomically
to ensure the count is always updated properly.

Thanks for Peter Butler <pbutler@pt.com> for his assistance in
diagnosing and fixing this problem.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <Hao.Xu@windriver.com>