tipc: Prevent unnecessary link reset during re-discovery of a failed node
commit f4a97ad5936c3bead4a1b14956bb68c8b59c7d1a from
http://tipc.cslab.ericsson.net/cgi-bin/gitweb.cgi?p=people/allan/tipc.git;a=shortlog;h=tipc1.7

Revises the handling of neighbor discovery messages to prevent TIPC
from resetting a working link to a newly rebooted node if the discovery
message arrives after the link endpoint has already started communicating
with its peer. Previously, the arrival of a discovery message with the
rebooted node's new signature would cause TIPC to reset its link endpoint;
now, TIPC simply updates its record of the node's signature as long as
the link endpoint is working and is communicating with its peer using
the media address specified in the discovery message.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <Hao.Xu@windriver.com>