tipc: Prevent null pointer error when removing a network subscription
commit 19e02f220375d7f33d2b59b0a4a2512f30773245 from
http://tipc.cslab.ericsson.net/cgi-bin/gitweb.cgi?p=people/allan/tipc.git;a=shortlog;h=tipc1.7

Prevents a null pointer dereference from occurring if a network
element subscription is triggered at the same time that the subscribing
port or publication is terminating the subscription. The problem arises
if the triggering routine asynchronously activates and deregisters the
subscription while deregistration is already underway. To avoid this
race condition the triggering routine now marks the subscription as
defunct (instead of deregistering it) and lets the subscribing port or
publication code terminate the subscription in the normal manner.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <Hao.Xu@windriver.com>