tipc: Prevent loss of fragmented messages over unicast links
commit 76a4ca0748b6d50082ecc080829ae1e3ca918f05
http://tipc.cslab.ericsson.net/cgi-bin/gitweb.cgi?p=people/allan/tipc.git;a=shortlog;h=tipc1.7

Modifies TIPC's unicast link endpoint so that an incoming fragmented
message is not lost if reassembly cannot begin because there currently
is no buffer big enough to hold the entire reassembled message. The link
endpoint now ignores the first fragment completely, which causes the
sending node to retransmit the first fragment so that reassembly can be
re-attempted.

Signed-off-by: Howard Xu <hao.xu@windriver.com>
Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <Hao.Xu@windriver.com>