tipc: Add flow control during TIPC name binding and unbinding
commit d9e85bf831ad0455eb6ed7d6af318fda8a919dae
http://tipc.cslab.ericsson.net/cgi-bin/gitweb.cgi?p=people/allan/tipc.git;a=shortlog;h=tipc1.7

Enhances TIPC to that name table update messages broadcast by a node
are not lost because of link congestion. (Previously, scenarios in
which applications published and/or withdrew names faster than TIPC
could distribute them would cause the broadcast link to congest and
attempt to reset itself, resulting in a kernel panic.)

Now, if a bind() operation cannot be completed because TIPC is unable
to distribute a newly published name due to link congestion the name
is removed from the node's local name table and the bind() operation
is suspended until the congestion condition clears, after which the
operation is automatically reattempted.

In contrast, if TIPC is unable to distribute a newly withdrawn name
due to link congestion during a bind() or close() operation the name
is still removed from the node's local name table and the operation
returns immediately; the undistributed publication is added to a new
data structure known as the "withdraw table". TIPC then re-attempts
distribution of all such withdrawn names on a periodic basis until
the withdraw table is empty.

The asymmetry in handling of the publish and withdraw cases exists
for several reasons. Firstly, it is desirable to suspend the publication
of new names when congestion occurs to prevent the backlog of undistributed
name table update messages from increasing indefinitely; such suspension
is not necessary in the withdraw case because an application cannot
withdraw more names than are currently published. Secondly, the suspension
of bind() avoids the complications that would result if a newly added
but not yet distributed name table entry was immediately withdrawn,
which would require TIPC to deal with the as-yet-unsent publication
message; an analogous situation cannot occur during a withdraw operation
since there is no way to undo the withdrawl of a name.

Signed-off-by: Ying Xue <ying.xue@windriver.com>
Signed-off-by: Hao Xu <hao.xu@windriver.com>
Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Integrated-by: Howard Xu <Hao.Xu@windriver.com>