kexec: fix vmcore for multi-cpus
When a crash kernel is invoked and the vmcore image captured, each
cpu is dumped into the vmcore image with it's data identified by
a header "reg/<PID>" where PID is the process ID that it was running.
Any cpu that is idle when a crash-kernel is invoked will have a PID=0

The problem occurs when the vmcore is loaded into gdb.  Gdb attempts
to create a thread for each cpu capture, using the pid (from the
"reg/<PID>" identifier) as an ID for the thread.  If multiple cpus
were idle at the time of the crash gdb attempts to create
multiple threads with the same id (0), which causes an exception within
gdb itself.

To correct this, all cpus that were idle (identified as running with
pid=0) will not have any data recorded in the vmcore image.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>