Kernel/LTTV event size differs for event udpv4_rcv_extended
This is because the inproper alignment for struct serialize_l4412228, this struct
has a 8-byte member.
For some 32-bit platforms, the cross-compiler gcc makes the struct serialize_l4412228
align to 8 bytes, this will lead to different size for event udpv4_rcv_extended
between kernel and lttv.
So force the struct serialize_l4412228 align to 4 bytes.

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>