Connect probes to KVM TRACE_EVENTS
This patch includes probes for the already defined KVM TRACE_EVENTS.
Also it exports the VMX exit strings during the statedump, that way
the analyzer can match the exit reason id with a more explicit
representation.
The new ltt_dump_enum_tables function aims to provide a generic way to
export strings in the traces when necessary.

[ edit: fixed trailing whitespaces ]

Signed-off-by: Julien Desfossez <julien.desfossez@polymtl.ca>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>