powerpc/lockdep: ignore MSR_PR check when CONFIG_PROVE_LOCKING is enabled
In order to fix a oops when CONFIG_IRQSOFF_TRACER is enabled,
the commit(tracing: fix oops for powerpc in irqsoff_tracer)
avoid to invoke trace_hardirqs_on before returning to user space.
But when CONFIG_PROVE_LOCKING is enabled, the trace_hardirqs_on will
use a different implementation. And the lockdep also require to
invoke it before returning to both user space or kernel context
with irq enabled. So ignore the MS_PR check when CONFIG_PROVE_LOCKING
is enabled.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>