powerpc: initialize saved frame pointer with a valid value
When a kernel thread clones a child kernel thread, it doesn't initialize
saved frame pointer, which may have a invalid value for calling macro
CALLER_ADDR1. This invalid value may lead to oops.

Here we initialize the child's saved frame pointer with it's parent's
sp, which is a valid and sane value for CALLER_ADDR1.

Signed-off-by: Liming Wang <liming.wang@windriver.com>