omap3/trace-clock: disable PMU when release omap3 trace-clock
When we startup oprofile, we will get the following error:
"oprofile: CPU0 PMNC still enabled when setup new event counter."
"failed to open the /dev/oprofile/buffer driver: Device or resource"
"busy"

This is because both oprofile and tracer clock will use PMU, and
unfortunately tracer clock occupied the PMU before the oprofile. The
original design is when call get_trace_clock(), it will enable PMU,
but when call put_trace_clock(), it will leave the PMU enabled. So the
oprofile will have no chance to hold the PMU and to startup
successfully.

To handle this problem, we need to disable PMU when we don't need
trace clock anymore.

Signed-off-by: Hui Wang <Hui.Wang@windriver.com>