ARM: sched_clock: allow init_sched_clock() to be called early
commit 211baa7016894c02fc18693e21ca479cd08ac0c0 upstream

sched_clock is supposed to be initialized early - in the recently added
init_early platform hook.  However, in doing so we end up calling
mod_timer() before the timer lists are initialized, resulting in an
oops.

Split the initialization in two - the part which the platform calls
early which starts things off.  The addition of the timer can be
delayed until after we have more of the kernel initialized - when the
normal time sources are initialized.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
[minor change to apply the patch]
Integrated-by: Junxiao Bi <junxiao.bi@windriver.com>