MIPS: Fix KDUMP calltrace by modifying ssleep to conditional ssleep
The calltrace is caused by bash thread scheduling in atomic context.
the preempt disable comes from the handling of the sysrq-c
(triggered by "echo c > /proc/sysrq-trigger") in __handle_sysrq(),
where it takes a spinlock before calling the sysrq handler. The handler
in turn calls die(), which, on MIPS, for some reason, sleeps for 5
seconds before calling panic(). A valid solution for this
would be to put the ssleep() call in die() within two checks.

Signed-off-by: Tian Le <le.tian@windriver.com>